test(vault): :recycle: share sync test fixtures via tests/sync-test-helpers.mjs
